requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=1.9.0
torchvision>=0.10.0
opencv-python==4.8.1.78
numpy<2
Pillow>=8.3.0
requests>=2.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
pandas>=1.3.0
PyYAML>=5.3.1
matplotlib>=3.2.2
seaborn>=0.11.0
tqdm>=4.64.0
scipy>=1.7.0
psutil>=5.8.0 